[db_persistence][db_importer] Use 64-bit integers for primary keys

+++ db_persistence/migrations/2018-02-18-184850_create_transactions/up.sql
@@ -1,8 +1,8 @@
 CREATE TABLE transactions (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     hash BYTEA NOT NULL,
     version INTEGER NOT NULL,
     lock_time INTEGER NOT NULL,
     creation_time INTEGER NOT NULL,
-    block_id INTEGER NOT NULL REFERENCES blocks (id)
+    block_id BIGINT NOT NULL REFERENCES blocks (id)
 )
