+++ db_persistence/migrations/2018-02-18-184909_create_addresses/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE addresses (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     hash BYTEA NOT NULL,
     base58check VARCHAR(36) NOT NULL,
-    output_id INTEGER NOT NULL REFERENCES outputs (id)
+    output_id BIGINT NOT NULL REFERENCES outputs (id)
 )
