+++ db_persistence/src/domain/newoutput.rs
@@ -7,13 +7,13 @@ pub struct NewOutput {
   pub output_index: i32,
   pub value: i64,
   pub script: Vec<u8>,
-  pub transaction_id: i32,
+  pub transaction_id: i64,
 }
 
 impl NewOutput {
   pub fn new(
     output: &blk_file_reader::Output,
-    transaction_id: i32,
+    transaction_id: i64,
   ) -> NewOutput {
     NewOutput {
       output_index: output.index as i32,
